Full Implementation on decommission-dss
Implemented the handler function and handler command of decommission-dss to free up the disks from its state and remove the DSS, as well as sending and receiving messages back from the disk to the server. Still need fixing on the DELETE message from User to Disk being sent, but not received.